// src/lib/widgets/text-editor/model/toolbar-options.enum.ts
export enum ToolbarOption {
  Heading1 = "Heading 1",
  Heading2 = "Heading 2",
  Heading3 = "Heading 3",
  Bold = "Bold",
  Italic = "Italic",
  Strikethrough = "Strikethrough",
  AlignLeft = "Align Left",
  AlignCenter = "Align Center",
  AlignRight = "Align Right",
  BulletList = "Bullet List",
  OrderedList = "Ordered List",
  Code = "Code",
  Highlight = "Highlight",
  Image = "Image",
}